# docs/openapi/control-plane.yaml
openapi: 3.1.0
info:
  title: nexus control-plane API
  version: "1.0.0"
  description: |
    Routing-plane administration surface. Authors tenancy and routing: accounts,
    workspaces, memberships, per-workspace auth-routes, and custom domains.

    Fail-closed: every data route requires a bearer credential that resolves to
    an INDIVIDUALLY IDENTIFIABLE admin token (`routing.admin_tokens`, minted via
    `POST /admin-tokens`; `ADMIN_TOKEN_PEPPER` keys verification). During the
    token migration only, the legacy shared `CONTROL_AUTH_TOKEN` is accepted iff
    `ADMIN_LEGACY_TOKEN_OK=true` (attributed `legacy-shared`, deprecation-warned).
    The process refuses to start without an explicit auth choice
    (`CONTROL_AUTH_DISABLED=true` is the dev opt-out). There is no workspace
    header. Every success body carries `"result":"ok"`.

    Admin action audit (admin-action-audit): every mutating call records one
    audit event ATOMICALLY with the mutation — an unrecorded mutation does not
    commit — and rejected authentications are recorded as denial events. Every
    mutating endpoint accepts the optional `x-acting-operator` header (see the
    ActingOperator parameter): recorded verbatim, marked asserted, never used
    for authorization. The ledger is queryable at `/audit/events` and
    exportable at `/audit/events/export`.

    Ids are SERVER-MINTED (`acct_<uuidv7>` / `ws_<uuidv7>`): creation returns the id
...
    `created:false` instead of minting a duplicate.

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /workspaces/{id}/auth-routes:
    get:
      tags: [auth-routes]
      summary: List auth-route rules
      operationId: listAuthRoutes
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Rules
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthRouteList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [auth-routes]
      summary: Upsert an auth-route rule
      operationId: upsertAuthRoute
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthRouteBody' }
      responses:
        '200':
          description: Stored rule (echoed)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthRouteUpsertResult' }
        '400':
          description: invalid_prefix, or requirements set with auth_required=false (requirements_need_auth)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Unknown workspace
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [auth-routes]
      summary: Delete an auth-route rule
      operationId: deleteAuthRoute
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthRouteDelete' }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthRouteDeleteResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: CONTROL_AUTH_TOKEN
  parameters:
...
    WorkspaceId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token (a denial event is recorded)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    AuthRouteBody:
      type: object
      properties:
        path_prefix:
          type: string
          description: Must start with '/'.
        auth_required: { type: boolean }
        requires_role: { type: [string, "null"] }
        requires_entitlement: { type: [string, "null"] }
        min_aal: { type: [integer, "null"], minimum: 0, maximum: 255 }
        account_scoped:
          type: boolean
          default: false
          description: When true, existence-hides non-members as 404 before the role gate.
      required: [path_prefix, auth_required]
    AuthRoute:
      type: object
      description: One stored rule.
      properties:
        path_prefix: { type: string }
        auth_required: { type: boolean }
        requires_role: { type: [string, "null"] }
        requires_entitlement: { type: [string, "null"] }
        min_aal: { type: [integer, "null"] }
        account_scoped: { type: boolean }
      required: [path_prefix, auth_required, account_scoped]
    AuthRouteList:
      type: object
      properties:
        workspace_id: { type: string }
        routes:
          type: array
          items: { $ref: '#/components/schemas/AuthRoute' }
      required: [workspace_id, routes]
    AuthRouteUpsertResult:
      type: object
      description: Echoes the stored rule.
      properties:
        result: { type: string, const: ok }
        workspace_id: { type: string }
        path_prefix: { type: string }
        auth_required: { type: boolean }
        requires_role: { type: [string, "null"] }
        requires_entitlement: { type: [string, "null"] }
        min_aal: { type: [integer, "null"] }
        account_scoped: { type: boolean }
      required: [result, workspace_id, path_prefix, auth_required, account_scoped]
    AuthRouteDelete:
      type: object
      properties:
        path_prefix: { type: string }
      required: [path_prefix]
    AuthRouteDeleteResult:
      type: object
      properties:
        result: { type: string, const: ok }
        workspace_id: { type: string }
        path_prefix: { type: string }
      required: [result, workspace_id, path_prefix]
...